requirements.txt
-e .
aws-cdk.aws_autoscaling
